Swift preview 6 updates.

+++ Travis/Deploy.sh
@@ -8,24 +8,31 @@
 
 RELEASE_NAME=""
 
-if [[ $TRAVIS_TAG == '' ]]; then
+if [[ -f ./Build/IORunnerInstaller ]]; then
 
-	if [[ $TRAVIS_IMAGE == 'Docker' ]]; then
+	if [[ $TRAVIS_TAG == '' ]]; then
 
-		RELEASE_NAME="IORunnerInstaller_${APP_VERSION}_${OS_RELEASE}_swift_3"
-	else
-		RELEASE_NAME="IORunnerInstaller_${APP_VERSION}_Darwin_swift_2.2"
+		if [[ $TRAVIS_IMAGE == 'Docker' ]]; then
+
+			RELEASE_NAME="IORunnerInstaller_${APP_VERSION}_${OS_RELEASE}_swift_3"
+		else
+			RELEASE_NAME="IORunnerInstaller_${APP_VERSION}_Darwin_swift_2.2"
+		fi
+
+		cp ./Build/IORunnerInstaller ./Build/${RELEASE_NAME}
+		zip ./Build/${RELEASE_NAME}.zip ./Build/${RELEASE_NAME}
+
+	curl \
+	-H "Accept: application/json" \
+	-H "Content-Type: application/zip" \
+	-H "X-IO-Travis-CommitID: ${TRAVIS_COMMIT}" \
+	-H "X-IO-Travis-FileName: ${RELEASE_NAME}.zip" \
+	--data-binary @./Build/${RELEASE_NAME}.zip \
+	"http://ilkerozcan.com.tr/iorunner/uploadbuild/"
+
 	fi
 
-	cp ./Build/IORunnerInstaller ./Build/${RELEASE_NAME}
-	zip ./Build/${RELEASE_NAME}.zip ./Build/${RELEASE_NAME}
-
-curl \
--H "Accept: application/json" \
--H "Content-Type: application/zip" \
--H "X-IO-Travis-CommitID: ${TRAVIS_COMMIT}" \
--H "X-IO-Travis-FileName: ${RELEASE_NAME}.zip" \
---data-binary @./Build/${RELEASE_NAME}.zip \
-"http://ilkerozcan.com.tr/iorunner/uploadbuild/"
-
+	exit(0)
+else
+	exit(1)
 fi
